style: section background color
remember the complicaitons of local scoping css can be an issue. there is diff output when building vs local deploy.
idk why.

TODO: need to find a way to local build and preview even when deploying to cloudflare. might be bettwe
to deploy to linode and learn how to balance with CF or bunny. With linode I would also have better control
over the server portion as well. i think?